+++ lib/components/table/column_filterer_directive.dart
@@ -1,0 +1,9 @@
+part of bs_table_directives;
+
+@Directive(selector: 'template[bs-column-filterer]')
+class BsColumnFiltererDirective {
+
+  BsColumnFiltererDirective(this.templateRef);
+
+  TemplateRef templateRef;
+}