+++ lib/components/table/table_directives.dart
@@ -11,12 +11,14 @@ import 'package:ng_bootstrap/components/input/input.dart';
 part 'column_directive.dart';
 part 'table_component.dart';
 part 'column_editor_directive.dart';
+part 'column_filterer_directive.dart';
 
 /// Directives needed for creating data table components
 const bsTableDirectives = const [
   BsTableComponent,
   BsColumnDirective,
   BsColumnEditorDirective,
+  BsColumnFiltererDirective,
   BsInput
 ];