+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.0
+
+- add `editable` attribute to `bs_table`
+- add `searchable` attribute and search fields row to `bs_table` (fix #85)
+- add `hideSelectColumn` attribute to `bs_table` (fix #92)
+
+## 1.0.3
+
+- upgrade dart-sdk version to `>=2.1.0-dev <3.0.0`
+
 ## 1.0.2
 
 - fix sorting by number or date (fix #117)